Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
opencv-python>=4.10.0
numpy>=1.26.0
pyserial>=3.5
setuptools>=65.0.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0

